feat: Replays use game ticks instead of run time.

+++ src/game/recordings.rs
@@ -15,7 +15,7 @@ pub enum EventKind {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Event {
     pub kind: EventKind,
-    pub at: f64,
+    pub at: usize,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -28,21 +28,21 @@ impl Recording {
         Recording { events: vec![] }
     }
 
-    pub fn push_action(&mut self, at: f64, action: actions::Action) {
+    pub fn push_action(&mut self, at: usize, action: actions::Action) {
         self.events.push(Event {
             kind: EventKind::Action(action),
             at,
         })
     }
 
-    pub fn push_piece(&mut self, at: f64, k: tetrominos::Kind) {
+    pub fn push_piece(&mut self, at: usize, k: tetrominos::Kind) {
         self.events.push(Event {
             kind: EventKind::PieceSpawned(k),
             at,
         })
     }
 
-    pub fn gameover(&mut self, at: f64) {
+    pub fn gameover(&mut self, at: usize) {
         self.events.push(Event {
             kind: EventKind::GameOver,
             at,
